refactor(console): api resource details page
